Change how the premium keys work, add a proper user premium key, add new premium pages

+++ loritta-serializable-commons/src/commonMain/kotlin/net/perfectdreams/loritta/serializable/ProfileSectionsResponse.kt
@@ -19,7 +19,7 @@ data class ProfileSectionsResponse(
 
     @Serializable
     data class DonationsWrapper(
-        val value: Double
+        val value: Int
     )
 
     @Serializable
